requirements.txt
pytube
ffmpeg-python
yt-dlp
pydub
openai-whisper
torch
dotenv
openai
